requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.9.0
azure-ai-textanalytics==5.3.0
langchain[openai]==0.0.345
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
sqlalchemy==2.0.23
pypdf2==3.0.1
python-docx==1.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic==2.5.0
aiofiles==23.2.1
sentence-transformers==2.2.2
aiosqlite==0.21.0
httpx==0.23.0
rank_bm25
